+++ src/sdk/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added `HasCommunicationSiteFeatures` methods on `IWeb` [jansenbe - Bert Jansen]
+- Web branding via `IBrandingManager` with support for getting and setting themes [jansenbe - Bert Jansen]
 
 ### Changed